Translating HTML to jade and making czar page fill in
The jade code is a lot shorter and maintainable because there's loops to
generate things. Server.js also encodes a few fake form submissions that
it passes to the czar page, which displays the information in an
awful-looking format (to be improved). Once the database is in place,
this should be a fully-functioning website.