Parsing form results in Node to improve sorting
Server.js now parses the result and makes a new object to send to the
jade which does a lot less parsing itself. This allows better sorting
because Node understands the results much better than the jade did.

This commit contains lots of commented out old code which I'll delete in
the next commit